Use whitespace as the lords of markdown intended
In an effort to stop the vicious cycle of trailing whitespace being removed
from a context where it's actually necessary and meaningful (to preserve
line breaks at the end of these project critical song lyrics), add a note
documenting that this is desired whitespace.

Future generations of whitespace removers, take heed.