Update the install destination of javascript to use vendor
It is not possible to configure the Rails application to not use the
asset pipeline in newer versions of Rails, so that setting has been
removed.

Additionally, we can install the javascript provided by the gem into the
vendor directory in the Rails app, so that the user can customize it if
needed.