Rename helper method to announcement_unread?
Previously this method was named announcement_hidden?

That name is an excellent description of the implementation, because the
helper method is checking the cookie with the announcement's key to
determine whether or not it's value has been set to "hidden", indicating
that the user has already seen the message.

However, we don't really care about this implementation detail, we care
about the high level concept of an announcement having been read or not,
hence this renaming of the method must commence.