Do not require `rails/all` in dummy app; just pull what we need
Sprockets version 4 assumes the presence of a `manifest.js` file which tells it
what assets to precompile. This is fine for normal rails apps, but for the
purposes of our test-suite-only dummy app, it doesn't make sense to have that
file.

Because we were requiring `rails/all` we were invoking the Sprockets config and
initialization process, which tried to find this file.

This changes switches the dummy app set up to only include the portions of rails
it needs to run the spec suite.